requirements.txt
python-telegram-bot[job-queue]==20.7
aiohttp==3.13.3
